aider: Refactored find_path function to correctly identify the connecting link between two search directions.
# Aider chat conversation:

USER: the path found ['https://en.wikipedia.org/wiki/Manchester_City_F.C.', 'https://en.wikipedia.org/wiki/ISBN_(identifier)', 'https://en.wikipedia.org/wiki/Fruit'] I tried going from manchester city to isbn to fruit but there was no link from ISBN to fruit on the isbn page

ASSISTANT: The issue you're experiencing is due to the fact that the current implementation of the `find_path` function in `server/crawler.py` is not correctly identifying the connecting link between the two search directions.

The current implementation is appending a link to the path as soon as it is discovered in the other direction's discovered set. However, this does not guarantee that this link is the actual connecting link between the two search directions. It could be that this link was discovered in the other direction's search but is not part of the shortest path from the start page to the finish page.

To fix this, we need to modify the `find_path` function to correctly identify the connecting link. This can be done by keeping track of the parent of each link in the search. Then, when a link is discovered in both directions, we can trace back from this link to the start page in the start direction's search and to the finish page in the finish direction's search to form the path.

Here are the *SEARCH/REPLACE* blocks:

server/crawler.py
```python
<<<<<<< SEARCH
    discovered_start = set()
    discovered_finish = set()
=======
    discovered_start = {start_page: None}
    discovered_finish = {finish_page: None}
>>>>>>> REPLACE
```

server/crawler.py
```python
<<<<<<< SEARCH
                    discovered_start.add(next_start)
=======
                    discovered_start[next_start] = vertex_start
>>>>>>> REPLACE
```

server/crawler.py
```python
<<<<<<< SEARCH
                    discovered_finish.add(next_finish)
=======
                    discovered_finish[next_finish] = vertex_finish
>>>>>>> REPLACE
```

server/crawler.py
```python
<<<<<<< SEARCH
                        path_start.append(next_start)
=======
                        # path_start.append(next_start)
>>>>>>> REPLACE
```

server/crawler.py
```python
<<<<<<< SEARCH
                        path_finish.append(next_finish)
=======
                        # path_finish.append(next_finish)
>>>>>>> REPLACE
```

server/crawler.py
```python
<<<<<<< SEARCH
                        path = path_start + path_finish[::-1]
=======
                        path_start = []
                        path_finish = []
                        current = next_start
                        while current is not None:
                            path_start.append(current)
                            current = discovered_start[current]
                        current = next_finish
                        while current is not None:
                            path_finish.append(current)
                            current = discovered_finish[current]
                        path = path_start[::-1] + path_finish[1:]
>>>>>>> REPLACE
```

These changes will ensure that the path from the start page to the finish page is correctly formed by tracing back from the connecting link to the start page in the start direction's search and to the finish page in the finish direction's search.